hash type fixes

+++ src/main/java/com/soapboxrace/core/bo/util/AccoladesFunc.java
@@ -227,7 +227,7 @@ public class AccoladesFunc {
 
 		LuckyDrawItem luckyDrawItem = new LuckyDrawItem();
 		luckyDrawItem.setDescription(desc);
-		luckyDrawItem.setHash(hash);
+		luckyDrawItem.setHash(hash.intValue());
 		luckyDrawItem.setIcon(icon);
 		luckyDrawItem.setRemainingUseCount(count);
 		luckyDrawItem.setResellPrice(price);
